Adjust stepping and CPU map to eschew rewiring.
Very simple really, GRBL does its step timing calculations and scheduling on one set of bitmasks,
therefore in the original setup, all GPIO pins for each operation must be on the same port.

However, it is simple enough to create the bitmask as 32-bits wide, let GRBL do all its math,
and then split it in half and write it to two separate ports as necessary.

+++ grbl/planner.h
@@ -62,7 +62,7 @@ typedef struct {
   uint32_t step_event_count; // The maximum step axis count and number of steps required to complete this block.
 
   #ifdef STM32F103C8
-    uint16_t direction_bits;    // The direction bit set for this block (refers to *_DIRECTION_BIT in config.h)
+    uint32_t direction_bits;    // The direction bit set for this block (refers to *_DIRECTION_BIT in config.h)
   #else
     uint8_t direction_bits;    // The direction bit set for this block (refers to *_DIRECTION_BIT in config.h)
   #endif
